# openapi/p4pa-workflow-hub.openapi.yaml
openapi: 3.0.1
info:
  title: p4pa-workflow-hub
  description: Api and Models
  version: 0.0.1
servers:
  - url: "http://localhost:8080/workflowhub"
paths:
...
  /workflow/classification/transfer/{orgId}:
    post:
      tags:
        - Classification
      operationId: transferClassification
      summary: Transfer classification
      description: It will Processes Transfer classification based on the provided parameters returning its workflowId
      parameters:
        - name: orgId
          in: path
          required: true
          description: the unique identifier of the organization.
          schema:
            type: integer
            format: int64
        - name: iuv
          in: query
          required: true
          description: the unique identifier of the payment (IUV).
          schema:
            type: string
        - name: iur
          in: query
          required: true
          description: the identifier of the receipt associated with the payment.
          schema:
            type: string
        - name: transferIndex
          in: query
          required: true
          description: the index of the transfer to be classified.
          schema:
            type: integer
      responses:
        '201':
          description: "Transfer classification workflow created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowCreatedDTO'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowErrorDTO'
  /workflow/classification/iud/{orgId}/notify-receipt:
    post:
      tags:
        - Classification
      operationId: iudClassificationByReceiptSignal
      summary: IUD classification
      description: It will Processes iud classification based on the provided parameters returning its workflowId
      parameters:
        - name: orgId
          in: path
          required: true
          description: the unique identifier of the organization.
          schema:
            type: integer
            format: int64
        - name: iud
          in: query
          required: true
          description: the unique identifier of the debt position (IUD).
          schema:
            type: string
        - name: iuv
          in: query
          required: true
          description: the unique identifier of the payment (IUV).
          schema:
            type: string
        - name: iur
          in: query
          required: true
          description: the identifier of the receipt associated with the payment.
          schema:
            type: string
        - name: transferIndexes
          in: query
          required: true
          description: the indexes of the transfer to be classified.
          schema:
            type: array
            items:
              type: integer
      responses:
        '201':
          description: "IUD classification workflow created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowCreatedDTO'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowErrorDTO'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowCreatedDTO'
              example:
                code: "WORKFLOW_BAD_REQUEST"
  /workflow/classification/iud/{orgId}/notify-payment-notification:
    post:
      tags:
        - Classification
      operationId: iudClassificationByPaymentNotificationSignal
      summary: IUD classification
      description: It will Processes iud classification based on the provided parameters returning its workflowId
      parameters:
        - name: orgId
          in: path
          required: true
          description: the unique identifier of the organization.
          schema:
            type: integer
            format: int64
        - name: iud
          in: query
          required: true
          description: the unique identifier of the debt position (IUD).
          schema:
            type: string
      responses:
        '201':
          description: "IUD classification workflow created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowCreatedDTO'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowErrorDTO'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowCreatedDTO'
              example:
                code: "WORKFLOW_BAD_REQUEST"
...
security:
  - BearerAuth: [ ]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    WorkflowCreatedDTO:
      type: object
      required:
        - workflowId
        - runId
      properties:
        workflowId:
          type: string
        runId:
          type: string
    WorkflowErrorDTO:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - WORKFLOW_INGESTION_FLOW_FILE_NOT_SUPPORTED
            - WORKFLOW_CONFLICT
            - WORKFLOW_NOT_FOUND
            - WORKFLOW_BAD_REQUEST
            - WORKFLOW_GENERIC_ERROR
            - WORKFLOW_INVALID_SYNC_DP_WF_EXECUTION_CONFIG
        message:
          type: string
...